requirements.txt
boto3 == 1.4.4
Django >= 1.11
django-extensions == 1.7.7
django-flatblocks == 0.9.3
django-tinymce-4 == 0.0.18
gunicorn == 19.6.0
djangorestframework == 3.5.4
djangorestframework-filters == 0.9.1
django-cors-headers-multi == 1.2.0
selenium == 3.0.2
behave-django == 0.4.1
coreapi == 2.2.4
django-guardian == 1.4.6
markdown == 2.6.8
django-filter == 1.0.1
lettuce_utils == 0.4.15
pandas == 0.19.2
PyPDF2 == 1.26.0
numpy == 1.12.1
scipy == 0.19.0
matplotlib == 2.0.0
utide == 0.1.0
django-grappelli == 2.9.1
django-filebrowser == 3.7.2
django-rest-knox == 3.0.0
prospector == 0.12.4
django-image-cropping==1.0.3
wand == 0.4.2
django-taggit == 0.20.2
django-sortedm2m == 1.3.3
easy-thumbnails==2.4.1